refactor: split compiled firewall matcher tests
