+++ crates/runner/mitm-addon/tests/firewall_helpers.py
@@ -27,6 +27,12 @@ def grant_all(firewalls, unknown_policy="deny"):
     return result
 
 
+def compile_firewalls_or_fail(firewalls):
+    compiled = matching.compile_firewalls(firewalls)
+    assert compiled is not None
+    return compiled
+
+
 def match_request_with_raw_firewalls(url, method, firewalls, network_policies=None):
     """Match raw firewall config through the production compiled matcher."""
     compiled_firewalls = matching.compile_firewalls(firewalls)
